Rename drawMultilineText to MultilineText
