+++ generators/DramaticCrossroads/generate.ts
@@ -1,6 +1,6 @@
 import { GeneratorFunction } from '../../types/GeneratorTypes';
 import { loadImage } from '../../utils/loadImage';
-import { MultilineText } from '../../utils/drawMultilineText';
+import { MultilineText } from '../../utils/MultilineText';
 
 import { buildAltText } from './buildAltText';
 import { FONT_SIZE } from './constants';
