+++ generators/EndlessCycle/generate.ts
@@ -1,5 +1,5 @@
 import { GeneratorFunction } from '../../types/GeneratorTypes';
-import { MultilineText } from '../../utils/drawMultilineText';
+import { MultilineText } from '../../utils/MultilineText';
 
 import { buildAltText } from './buildAltText';
 import { FONT_SIZE } from './constants';
